cleaned up test_verification_module with conftest.py
