+++ tests/test_verification_module/test_nnenum_module.py
@@ -1,35 +1,31 @@
 import pytest
 import subprocess
 from unittest.mock import patch
-import torch
-from robustness_experiment_box.verification_module.nnenum_module import NnenumModule
+from torch import load
+from result import Err, Ok
 from robustness_experiment_box.database.verification_context import VerificationContext
-from robustness_experiment_box.database.vnnlib_property import VNNLibProperty
-from robustness_experiment_box.database.network import Network
-from robustness_experiment_box.database.dataset.data_point import DataPoint
+
 
 from autoverify.verifier.verification_result import CompleteVerificationData
-import numpy as np
 
 from robustness_experiment_box.verification_module.property_generator.one2any_property_generator import One2AnyPropertyGenerator
 from robustness_experiment_box.verification_module.property_generator.one2one_property_generator import One2OnePropertyGenerator
 
 
-@pytest.fixture
-def network():
-    return Network("/path/to/network")
-
-@pytest.fixture
-def datapoint():
-    return DataPoint("1", 0, torch.tensor([0.1, 0.2, 0.3]))  
 
 @pytest.fixture
 def verification_context(network, datapoint, tmp_path, property_generator):
-    return VerificationContext(network, datapoint, tmp_path, property_generator)
+    return VerificationContext(network=network, data_point=datapoint, tmp_path=tmp_path, property_generator=property_generator)
+
 
 @pytest.fixture
-def nnenum_module():
-    return NnenumModule(timeout=60.0)
+def result(datapoint):
+    data = """
+   (X_3 0.1)
+   (Y_0 0.2)
+    """
+    return Ok(CompleteVerificationData(result="SAT", counter_example=data, took =10))
+
 
 @patch('subprocess.run')
 @pytest.mark.parametrize("property_generator", [One2AnyPropertyGenerator(), One2OnePropertyGenerator(target_class=0)])
